examples/scaffolding/mcp/python_interpreter/pyproject.toml
[project]
name = "python-interpreter-mcp"
version = "0.1.0"
description = "MCP server for executing Python code in a sandboxed environment"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
    "starlette>=0.38.0",
]
